+++ crates/storage/trie_db/rocksdb_locked.rs
@@ -3,7 +3,7 @@ use ethrex_trie::{Nibbles, TrieDB, error::TrieError};
 use rocksdb::{DBWithThreadMode, MultiThreaded, SnapshotWithThreadMode};
 use std::sync::Arc;
 
-use crate::{store_db::rocksdb::CF_FLATKEYVALUE, trie_db::layering::apply_prefix};
+use crate::trie_db::layering::apply_prefix;
 
 /// RocksDB locked implementation for the TrieDB trait, read-only with consistent snapshot.
 pub struct RocksDBLockedTrieDB {
@@ -23,7 +23,8 @@ pub struct RocksDBLockedTrieDB {
 impl RocksDBLockedTrieDB {
     pub fn new(
         db: Arc<DBWithThreadMode<MultiThreaded>>,
-        cf_name: &str,
+        trie_cf_name: &str,
+        fkv_cf_name: &str,
         address_prefix: Option<H256>,
         last_written: Vec<u8>,
     ) -> Result<Self, TrieError> {
@@ -31,12 +32,12 @@ impl RocksDBLockedTrieDB {
         let db = Box::leak(Box::new(db));
 
         // Verify column family exists
-        let cf = db.cf_handle(cf_name).ok_or_else(|| {
-            TrieError::DbError(anyhow::anyhow!("Column family not found: {}", cf_name))
+        let cf = db.cf_handle(trie_cf_name).ok_or_else(|| {
+            TrieError::DbError(anyhow::anyhow!("Column family not found: {}", trie_cf_name))
         })?;
         // Verify column family exists
-        let cf_flatkeyvalue = db.cf_handle(CF_FLATKEYVALUE).ok_or_else(|| {
-            TrieError::DbError(anyhow::anyhow!("Column family not found: {}", cf_name))
+        let cf_flatkeyvalue = db.cf_handle(fkv_cf_name).ok_or_else(|| {
+            TrieError::DbError(anyhow::anyhow!("Column family not found: {}", trie_cf_name))
         })?;
 
         let last_computed_flatkeyvalue = Nibbles::from_hex(last_written);